feat: extract mouse/keyboard actions
